% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_accel.R
\name{age_accel}
\alias{age_accel}
\title{Calculate Age Acceleration}
\usage{
age_accel(Age, BA, Sex)
}
\arguments{
\item{Age}{Numeric vector (chronological age)}

\item{BA}{Numeric vector (biological age)}

\item{Sex}{Character vector ("Men", "Women")}
}
\value{
Numeric vector of residuals
}
\description{
Computes biological age acceleration as residuals from
a linear model of BA ~ Age, stratified by sex.
}
