% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnthropoAge-package.R
\docType{package}
\name{AnthropoAge-package}
\alias{AnthropoAge}
\alias{AnthropoAge-package}
\title{AnthropoAge: Calculate second-generation biological aging measures using AnthropoAge and PhenoAge}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Omar Yaxmehen Bello-Chavolla \email{oyaxbell@yahoo.com.mx} (\href{https://orcid.org/0000-0003-3093-937X}{ORCID})

Other contributors:
\itemize{
  \item Carlo Alberto Fermin-Martínez \email{carlosfermin38@gmail.com} (\href{https://orcid.org/0000-0001-5627-8851}{ORCID}) [contributor]
}

}
\keyword{internal}
