% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoage.R
\name{phenoage}
\alias{phenoage}
\title{Calculate PhenoAge}
\usage{
phenoage(Age, CRP, Lymph, WBC, Glu, RCDW, Alb, Cr, MCV, AP)
}
\arguments{
\item{Age}{Numeric vector}

\item{CRP}{C-reactive protein}

\item{Lymph}{Lymphocyte percentage}

\item{WBC}{White blood cell count}

\item{Glu}{Glucose (mg/dL)}

\item{RCDW}{Red cell distribution width}

\item{Alb}{Albumin}

\item{Cr}{Creatinine (mg/dL)}

\item{MCV}{Mean corpuscular volume}

\item{AP}{Alkaline phosphatase}
}
\value{
Numeric vector
}
\description{
Computes Phenotypic Age (PhenoAge) using clinical biomarkers.
}
