% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_anthropoage.R
\name{s_anthropoage}
\alias{s_anthropoage}
\title{Calculate Simplified AnthropoAge (S-AnthropoAge)}
\usage{
s_anthropoage(Age, Sex, Height, Weight, Waist, Ethnicity)
}
\arguments{
\item{Age}{Numeric vector}

\item{Sex}{Character vector ("Men", "Women")}

\item{Height}{Numeric (meters)}

\item{Weight}{Numeric (kg)}

\item{Waist}{Numeric}

\item{Ethnicity}{Character vector ("White", "Black", "Mexican-American", "Other")}
}
\value{
Numeric vector
}
\description{
Computes S-AnthropoAge as a proxy of biological age using
anthropometric variables, age, sex and ethnicity.
}
