% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropoage.R
\name{anthropoage}
\alias{anthropoage}
\title{Calculate AnthropoAge}
\usage{
anthropoage(
  Age,
  Sex,
  Ethnicity,
  Height,
  Weight,
  Waist,
  Subscapular_skinfold = 1,
  Triceps_skinfold = 1,
  Thigh_circumference = 1,
  Arm_circumference = 1
)
}
\arguments{
\item{Age}{Numeric vector}

\item{Sex}{Character vector ("Men", "Women")}

\item{Ethnicity}{Character vector ("White", "Black", "Mexican-American", "Other")}

\item{Height}{Numeric (meters)}

\item{Weight}{Numeric (kg)}

\item{Waist}{Numeric}

\item{Subscapular_skinfold}{Numeric (optional)}

\item{Triceps_skinfold}{Numeric (optional)}

\item{Thigh_circumference}{Numeric (optional)}

\item{Arm_circumference}{Numeric (optional)}
}
\value{
Numeric vector of AnthropoAge
}
\description{
Computes AnthropoAge as a proxy of biological age using
anthropometric variables, age, sex and ethnicity.
}
