% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnthropoAge-package.R
\docType{package}
\name{AnthropoAgeR-package}
\alias{AnthropoAgeR}
\alias{AnthropoAgeR-package}
\title{AnthropoAgeR: Calculate Second Generation Biological Aging Measures Using AnthropoAge and PhenoAge}
\description{
Description: Provides tools to estimate second-generation biological age biomarkers derived from anthropometric and clinical data, including AnthropoAge, S-AnthropoAge, and PhenoAge. The package also computes sex-stratified age acceleration measures (AnthropoAgeAccel and PhenoAgeAccel) based on published models and supports both individual-level and cohort-level calculations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oyaxbell/AnthropoAgeR}
  \item Report bugs at \url{https://github.com/oyaxbell/AnthropoAgeR/issues}
}

}
\author{
\strong{Maintainer}: Omar Yaxmehen Bello-Chavolla \email{oyaxbell@yahoo.com.mx} (\href{https://orcid.org/0000-0003-3093-937X}{ORCID})

Other contributors:
\itemize{
  \item Carlo Alberto Fermin-Martínez \email{carlosfermin38@gmail.com} (\href{https://orcid.org/0000-0001-5627-8851}{ORCID}) [contributor]
}

}
\keyword{internal}
