% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_anthropoage_fast.R
\name{s_anthropoage_fast}
\alias{s_anthropoage_fast}
\title{Fast computation of S-AnthropoAge}
\usage{
s_anthropoage_fast(Age, Sex, Height, Weight, Waist, Ethnicity)
}
\arguments{
\item{Age}{Numeric vector of age in years}

\item{Sex}{Character vector ("Men", "Women")}

\item{Height}{Numeric vector (meters)}

\item{Weight}{Numeric vector (kg)}

\item{Waist}{Numeric vector (cm)}

\item{Ethnicity}{Character vector ("White", "Black", "Mexican-American", "Other")}
}
\value{
Numeric vector of S-AnthropoAge
}
\description{
Computes S-AnthropoAge as a proxy of biological age using
anthropometric variables, age, sex and ethnicity using a vectorized approach.
}
